Profile.tsxのロジック部分をカスタムフック化
